Fix build for Windows 10.

+++ belr/src/belr.cc
@@ -485,7 +485,7 @@ void Grammar::optimize(){
 
 
 int Grammar::getNumRules() const{
-	return mRules.size();
+	return (int)mRules.size();
 }
